feat: fix loading issues with ktx

+++ source/ref_base/r_texture_buf.c
@@ -290,13 +290,13 @@ void T_SwizzleInplace(struct texture_buf_s* tex, enum texture_logical_channel_e*
 			for( size_t row = 0; row < logicalHeight; row++ ) {
 				for( size_t column = 0; column < logicalWidth; column++ ) {
 					uint8_t * const block = &tex->buffer[(tex->rowPitch * row) + ( column * RT_BlockSize(tex->def))];
-					// save the values 
-					for(uint8_t c  = 0; c < tex->def->fixed_8.numChannels; c++) {
+					// save the values
+					for( size_t c = 0; c < tex->def->fixed_8.numChannels; c++ ) {
 						assert(tex->def->fixed_8.channels[c] < R_LOGICAL_C_MAX);
-						values[tex->def->fixed_8.channels[c]] = *(block + c); 
+						values[tex->def->fixed_8.channels[c]] = *(block + c);
 					}
 					// write the new values to the block
-					for(uint8_t c  = 0; c < tex->def->fixed_8.numChannels; c++) {
+					for(size_t c  = 0; c < tex->def->fixed_8.numChannels; c++) {
 						assert(channels[c] < R_LOGICAL_C_MAX);
 						(*(block + c)) = values[channels[c]];
 					}
@@ -310,12 +310,12 @@ void T_SwizzleInplace(struct texture_buf_s* tex, enum texture_logical_channel_e*
 				for( size_t column = 0; column < logicalWidth; column++ ) {
 					uint16_t * const block = (uint16_t*)(&tex->buffer[(tex->rowPitch * row) + ( column * RT_BlockSize(tex->def))]);
 					// save the values 
-					for(uint16_t c  = 0; c < tex->def->fixed_8.numChannels; c++) {
+					for(size_t  c  = 0; c < tex->def->fixed_8.numChannels; c++) {
 						assert(tex->def->fixed_16.channels[c] < R_LOGICAL_C_MAX);
 						values[tex->def->fixed_16.channels[c]] = *(block + c); 
 					}
 					// write the new values to the block
-					for(uint8_t c  = 0; c < tex->def->fixed_8.numChannels; c++) {
+					for(size_t c  = 0; c < tex->def->fixed_8.numChannels; c++) {
 						assert(channels[c] < R_LOGICAL_C_MAX);
 						(*(block + c)) = values[channels[c]];
 					}
